Revert "Updated files to use LTC instead of DOGE. Updated to use any tip amount as float w/ 8 point precision, as it was previously rounded to the nearest int."
This reverts commit 1da28c54097ca25ea133ba9a2d0a8a8d05327a98.

+++ utils.py
@@ -47,7 +47,7 @@ def get_coin_value(balance, format=2):
                 jc_currency = c_currency.json()
                 bot_logger.logger.info('value is $%s' % str(jc_currency['Data'][0]['Price']))
                 usd_currency = float(str("{0:." + str(format) + "f}").format(
-                    float(balance) * float(jc_currency['Data'][0]['Price'])))
+                    int(balance) * float(jc_currency['Data'][0]['Price'])))
                 return usd_currency
             except:
                 traceback.print_exc()
